modified:   .vscode/c_cpp_properties.json 	modified:   .vscode/keil-assistant.log 	modified:   Hardware/BlueTooth.c 	modified:   Objects/test.bin 	modified:   Objects/test.elf 	modified:   Task/Task_BlueTooth.c 	modified:   Task/Task_Config.h 	modified:   Task/Task_Motor_Control.c 	modified:   Task/Task_NRF24L001.c 	modified:   Task/Task_OLED.c 	modified:   Task/Task_Run.c 	modified:   test.uvguix.linwei 	modified:   test.uvoptx

+++ Task/Task_Config.h
@@ -56,8 +56,8 @@ typedef struct {
 
 /*速度设置结构体*/
 typedef struct {
-	float target_speed;
-	float turn_speed;
+	int8_t target_speed;
+	int8_t turn_speed;
 }Target_Speed_t;
 
 /*PID结构体*/
